+++ obj/Debug/netstandard2.1/HeavyItemSCPs.AssemblyInfo.cs
@@ -78,7 +78,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ProjectSCP.HeavyItemSCPs")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.4.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.4.0+f8e219a8b2d06cf4c6ef34283769d10fa76f1c00")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.4.0+b981a6166091ae8719951ae7412218da104ffbbb")]
 [assembly: System.Reflection.AssemblyProductAttribute("HeavyItemSCPs")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ProjectSCP.HeavyItemSCPs")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.4.0.0")]